ID-428 working on invite user specs
